+++ ax3l/app/snakelab/prompts/NoDupes.py
@@ -3,6 +3,6 @@
 from ax3l.app.Prompt import Prompt
 
 
-class NoDupesSingle(Prompt):
+class NoDupes(Prompt):
     def __init__(self, reason: str):
         super().__init__(f"{reason}. Choose a different value and submit it again.")